Rename casebook_actionbutton_helper.
WIP refactor node_decorator.
